feat(dashboard): update workflow tabs and add child thread variables
