+++ dashboard/src/components/diagram/left-sidebar/index.tsx
@@ -6,12 +6,12 @@ import { LeftSidebarTabId } from '@/types'
 import { LHStatus, WfRun, WfSpec } from 'littlehorse-client/proto'
 import { useEffect, useState } from 'react'
 import WfRunTab from './tab-content/wf-run-tab'
-import WfSpecTab from './tab-content/wf-spec-tab'
+import WorkflowTab from './tab-content/workflow-tab'
 import LeftSidebarTabs from './tabs'
 import ActionButton from './action-button'
 
 const tabDescriptions: Record<LeftSidebarTabId, string> = {
-  WfSpec: 'Workflow Specification',
+  Workflow: 'WfSpec & WfRun',
   WfRuns: 'Workflow Runs',
   ScheduledWfRuns: 'Scheduled Workflow Runs',
 }
@@ -24,14 +24,14 @@ interface LeftSidebarProps {
 }
 
 export default function LeftSidebar({ wfSpec, wfRun }: LeftSidebarProps) {
-  const [activeTab, setActiveTab] = useState<LeftSidebarTabId>('WfSpec')
+  const [activeTab, setActiveTab] = useState<LeftSidebarTabId>('Workflow')
   const [sidebarState, setSidebarState] = useState<SidebarState>('normal')
 
   // Auto-expand sidebar when tab is WfRuns or ScheduledWfRuns
   useEffect(() => {
     if (activeTab === 'WfRuns' || activeTab === 'ScheduledWfRuns') {
       setSidebarState(prev => (prev === 'hidden' ? 'hidden' : 'expanded'))
-    } else if (activeTab === 'WfSpec') {
+    } else if (activeTab === 'Workflow') {
       setSidebarState(prev => (prev === 'hidden' ? 'hidden' : 'normal'))
     }
   }, [activeTab])
@@ -92,13 +92,12 @@ export default function LeftSidebar({ wfSpec, wfRun }: LeftSidebarProps) {
             <span className="text-sm text-[#656565]">{`v${wfSpec?.id?.majorVersion}.${wfSpec?.id?.revision}`}</span>
             <Badge
               variant="outline"
-              className={`ml-3 ${
-                wfRun?.status === LHStatus.COMPLETED
-                  ? 'bg-emerald-100 text-emerald-600'
-                  : wfRun?.status === LHStatus.RUNNING
-                    ? 'bg-blue-100 text-blue-600'
-                    : 'bg-emerald-100 text-emerald-600'
-              } hover:bg-[#c5d0ff]/90`}
+              className={`ml-3 ${wfRun?.status === LHStatus.COMPLETED
+                ? 'bg-emerald-100 text-emerald-600'
+                : wfRun?.status === LHStatus.RUNNING
+                  ? 'bg-blue-100 text-blue-600'
+                  : 'bg-emerald-100 text-emerald-600'
+                } hover:bg-[#c5d0ff]/90`}
             >
               {wfRun?.status ?? wfSpec?.status}
             </Badge>
@@ -114,7 +113,7 @@ export default function LeftSidebar({ wfSpec, wfRun }: LeftSidebarProps) {
             <span className="font-medium">{tabDescriptions[activeTab]}</span>
           </div>
           <div className="flex-1 overflow-auto">
-            {activeTab === 'WfSpec' && <WfSpecTab wfSpec={wfSpec} wfRun={wfRun} />}
+            {activeTab === 'Workflow' && <WorkflowTab wfSpec={wfSpec} wfRun={wfRun} />}
             {activeTab === 'WfRuns' && <WfRunTab />}
             {activeTab === 'ScheduledWfRuns' && <ScheduledWfRunsTab />}
           </div>
